PhotoOrganizer/requirements.txt
ImageHash==4.3.2
iniconfig==2.1.0
numpy==2.3.1
packaging==25.0
piexif==1.1.3
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
PyWavelets==1.8.0
scipy==1.16.0
tqdm==4.67.1
